update: move renderer to a component
