+++ packages/@dcl/inspector/public/index.html
@@ -11,9 +11,6 @@
 <body class="with-sidebar">
   <noscript>You need to enable JavaScript to run this app.</noscript>
   <div id="root"></div>
-  <div id="main-editor">
-    <canvas id="renderer" touch-action="none"></canvas>
-  </div>
   <link href="bundle.css" rel="stylesheet" />
   <script src="bundle.js"></script>
 </body>
