+++ packages/@dcl/inspector/src/index.css
@@ -25,6 +25,7 @@ code {
 }
 
 #root {
+  display: flex;
   flex-basis: 20rem;
   flex-grow: 1;
   overflow: scroll;
